# Limit allocations in concurrent C# impl