We were allocating 3.5 GB of arrays. I'm seeing an 8% improvement, but it should ideally be measured on the benchmark machine because I wouldn't be too surprised if the OS was able to give out fresh zero initialized pages faster than we can zero them manually -  and this is not _too many_ individual allocations to garbage collect.