@jinyus 2023-10-19
Main branch:
```
C# Concurrent (JIT) |    23.23 ms    |    161.49 ms    |    1.29 s 
C# Concurrent (AOT) |    16.67 ms    |    169.15 ms    |    1.41 s
```

This PR:
```
C# Concurrent (JIT):

        Benchmark 1: ./bin/release/net8.0/jit/related
        Processing time (w/o IO): 10.8715ms
        Processing time (w/o IO): 10.6414ms
        Processing time (w/o IO): 10.7252ms
        Processing time (w/o IO): 10.9707ms
        Processing time (w/o IO): 10.9424ms
        Processing time (w/o IO): 10.5698ms
        Processing time (w/o IO): 10.5083ms
        Processing time (w/o IO): 10.524ms
        Processing time (w/o IO): 10.8871ms
        Processing time (w/o IO): 10.8364ms
        Processing time (w/o IO): 11.1831ms
        Processing time (w/o IO): 10.414ms
        Processing time (w/o IO): 11.0252ms
          Time (mean ± σ):     188.5 ms ±   5.8 ms    [User: 198.2 ms, System: 35.8 ms]
          Range (min … max):   182.0 ms … 198.7 ms    10 runs
         
C# Concurrent (JIT):

        Benchmark 1: ./bin/release/net8.0/jit/related
        Processing time (w/o IO): 125.2667ms
        Processing time (w/o IO): 122.7871ms
        Processing time (w/o IO): 122.6332ms
          Time (mean ± σ):     577.9 ms ±   6.0 ms    [User: 1196.4 ms, System: 69.7 ms]
          Range (min … max):   573.7 ms … 582.2 ms    2 runs
         
C# Concurrent (JIT):

        Benchmark 1: ./bin/release/net8.0/jit/related
        Processing time (w/o IO): 1009.469ms
        Processing time (w/o IO): 1013.3985ms
        Processing time (w/o IO): 1011.2509ms
          Time (mean ± σ):      3.003 s ±  0.111 s    [User: 8.663 s, System: 0.172 s]
          Range (min … max):    2.924 s …  3.081 s    2 runs


C# Concurrent (AOT):

        Benchmark 1: ./bin/release/net8.0/aot/related
        Processing time (w/o IO): 10.679ms
        Processing time (w/o IO): 10.5957ms
        Processing time (w/o IO): 10.6472ms
        Processing time (w/o IO): 11.0456ms
        Processing time (w/o IO): 11.0136ms
        Processing time (w/o IO): 10.66ms
        Processing time (w/o IO): 10.8861ms
        Processing time (w/o IO): 10.7932ms
        Processing time (w/o IO): 10.6774ms
        Processing time (w/o IO): 10.7497ms
        Processing time (w/o IO): 10.6758ms
        Processing time (w/o IO): 10.8635ms
        Processing time (w/o IO): 10.785ms
          Time (mean ± σ):      73.2 ms ±   2.9 ms    [User: 128.2 ms, System: 25.4 ms]
          Range (min … max):    70.2 ms …  80.4 ms    10 runs
         
C# Concurrent (AOT):

        Benchmark 1: ./bin/release/net8.0/aot/related
        Processing time (w/o IO): 136.1051ms
        Processing time (w/o IO): 135.554ms
        Processing time (w/o IO): 135.5004ms
          Time (mean ± σ):     439.6 ms ±   1.8 ms    [User: 1186.8 ms, System: 55.5 ms]
          Range (min … max):   438.4 ms … 440.9 ms    2 runs
         
C# Concurrent (AOT):

        Benchmark 1: ./bin/release/net8.0/aot/related
        Processing time (w/o IO): 1132.723ms
        Processing time (w/o IO): 1134.3475ms
        Processing time (w/o IO): 1131.4183ms
          Time (mean ± σ):      2.807 s ±  0.001 s    [User: 9.357 s, System: 0.169 s]
          Range (min … max):    2.806 s …  2.808 s    2 runs
         

```